# lib/api-zod/src/generated/types/aiRouteRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIOS — Personal AI Operating System API
 * OpenAPI spec version: 0.1.0
 */
import type { AiRouteRequestTaskType } from './aiRouteRequestTaskType';

export interface AiRouteRequest {
  taskType: AiRouteRequestTaskType;
  /**
     * @minimum 0
     * @maximum 1
     */
  complexity?: number;
}