# lib/api-zod/src/generated/types/aiRouteResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIOS — Personal AI Operating System API
 * OpenAPI spec version: 0.1.0
 */
import type { AiRouteResponseTier } from './aiRouteResponseTier';

export interface AiRouteResponse {
  modelId: string;
  modelName: string;
  tier: AiRouteResponseTier;
  reason: string;
  estimatedTokens?: number;
  estimatedCostUsd?: number;
}