# lib/api-zod/src/generated/types/conversationMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIOS — Personal AI Operating System API
 * OpenAPI spec version: 0.1.0
 */
import type { ConversationMessageRole } from './conversationMessageRole';

export interface ConversationMessage {
  id: string;
  conversationId: string;
  projectId: string;
  role: ConversationMessageRole;
  content: string;
  modelId?: string;
  tokensUsed?: number;
  createdAt: Date;
}