# lib/api-zod/src/generated/types/createVersionRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIOS — Personal AI Operating System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateVersionRequestBumpType } from './createVersionRequestBumpType';

export interface CreateVersionRequest {
  changelog: string;
  bumpType: CreateVersionRequestBumpType;
}